Update for German Localizable.strings
The original German (machine?) translation wasn’t very good and only just intelligible. I tried to enhance it as good as I could in a first attempt. (I am a German native speaker.) I sticked with "Loading" as a) it is well known as a term for the process of downloading data in German and "Laden" does not convey that it is something that is going on right now and b) the "LOADING" string obviously is used as name of the app in the Options > About menu. (The actual fix would be to differentiate here between strings for the process of loading and the name Loading.) I haven’t checked where "DOWNLOAD" is used, but I guess it is for the download of an update, so "Laden" is okay, I guess, at least in contemporary colloquial German.
(Finally, may I add that the German app description on the app’s homepage is quite horrible as well?)